nu är koden lite finare
